Permit files with multiple newlines between blocks
The WebVTT file structure spec[1] specifies that style, comment and cue blocks
are to be separated from each other by one or more WebVTT line terminators.
The following should therefore be considered valid input:

    WEBVTT

    00:00:29.000 --> 00:00:31.000 line:75%
    English subtitle 15 -Forced- (00:00:27.000)
    line:75%

    00:00:31.000 --> 00:00:33.000 align:start line:0%
    English subtitle 16 -Unforced- (00:00:31.000)
    align:start line:0%

The parser previously split the input content on "\n\n", which resulted in
empty strings being interpreted as cue blocks, causing an exception to be
raised when Cue.parse was invoked with them.

This alters the content chunking to split on the presence of two or more
linefeed characters instead, so any extra blank lines will be consumed without
parsing.

[1] https://w3c.github.io/webvtt/#file-structure